+++ packages/test/jest.config.ts
@@ -13,7 +13,7 @@ module.exports = {
     ],
     '^.+\\.[t|j]sx?$': 'babel-jest',
   },
-  setupFilesAfterEnv: ['<rootDir>/__mocks__/consoleMock.ts'],
+  setupFilesAfterEnv: ['<rootDir>/__mocks__/consoleMock.ts', '<rootDir>/__mocks__/fetchMock.ts'],
   moduleNameMapper: {
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
       '<rootDir>/__mocks__/fileMock.ts',